Mindex Coding challenge
Persistence for Compensation is not fully implemented. File writing only works to the temporary file compensation_database.json in the build package. Persisntence works if this file is opened and watched while instances are killed/restarted. File in src/main/resources (or anywhere for that matter) would generate an error that the path is invalid, regardless of where in the file structure the .json file lived. I left the incomplete file writing techniques in the code.

Unit tests provided by Mindex would not run successfully. I added Unit tests of my own which fail for the same reason as the provided tests.